+++ src/features/standard/custom-left-sidebar/custom-left-sidebar.ts
@@ -2,11 +2,16 @@ import css from '@src/utils/css-utils.module.css';
 import { createFragmentApp } from '@src/utils/vue-fragment-app';
 import SidebarButtons from './SidebarButtons.vue';
 import { applyCssRule } from '@src/infrastructure/prun-ui/refined-prun-css';
+import { refAttributeValue } from '@src/utils/reactive-dom';
 
 function init() {
   applyCssRule('#TOUR_TARGET_SIDEBAR_LEFT_02', css.hidden);
   subscribe($$(document, C.Frame.sidebar), sidebar => {
-    createFragmentApp(SidebarButtons).appendTo(sidebar);
+    const com = _$$(sidebar, C.Frame.toggle).find(x => x.textContent === 'COM');
+    const comIndicator = com ? _$(com, C.Frame.toggleIndicator) : undefined;
+    const comClass = comIndicator ? refAttributeValue(comIndicator, 'class') : ref(undefined);
+    const comPulse = computed(() => comClass.value?.includes(C.Frame.toggleIndicatorPulseActive));
+    createFragmentApp(SidebarButtons, reactive({ comPulse })).appendTo(sidebar);
   });
 }